Movie details unifished demo

+++ backend/routes/movieRoutes.js
@@ -1,6 +1,6 @@
 import express from 'express'
 import Movie from '../models/Movie.js'
-// import checkJwt from '../jwt.js'
+import checkJwt from '../jwt.js'
 
 const router = express.Router()
 
@@ -18,12 +18,26 @@ const router = express.Router()
  *             required:
  *               - title
  *               - director
+ *               - userId
+ *               - movieId
  *             properties:
  *               title:
+ *                 type: string
+ *               userId:
  *                 type: string
  *               director:
  *                 type: string
  *               year:
+ *                 type: number
+ *               ImdbRate:
+ *                 type: number
+ *               language:
+ *                 type: array
+ *                 items:
+ *                   type: string
+ *               movieId:
+ *                 type: number
+ *               personalRating:
  *                 type: number
  *     responses:
  *       201:
@@ -31,18 +45,16 @@ const router = express.Router()
  *       400:
  *         description: Bad request
  */
-router.post(
-  '/api/movies',
-  /*checkJwt,*/ async (req, res) => {
-    try {
-      const movie = new Movie({ ...req.body })
-      await movie.save()
-      res.status(201).send(movie)
-    } catch (error) {
-      res.status(400).send({ error: error.message })
-    }
-  },
-)
+router.post('/api/movies', checkJwt, async (req, res) => {
+  try {
+    const userId = req.auth.sub
+    const movie = new Movie({ ...req.body, userId })
+    await movie.save()
+    res.status(201).send(movie)
+  } catch (error) {
+    res.status(400).send({ error: error.message })
+  }
+})
 
 /**
  * @swagger
@@ -53,7 +65,7 @@ router.post(
  *       - in: path
  *         name: id
  *         schema:
- *           type: string
+ *           type: number
  *         required: true
  *         description: The movie ID
  *     requestBody:
@@ -69,6 +81,14 @@ router.post(
  *                 type: string
  *               year:
  *                 type: number
+ *               ImdbRate:
+ *                 type: number
+ *               language:
+ *                 type: array
+ *                 items:
+ *                   type: string
+ *               personalRating:
+ *                 type: number
  *     responses:
  *       200:
  *         description: Returns the updated movie
@@ -77,20 +97,20 @@ router.post(
  *       404:
  *         description: Movie not found
  */
-router.patch(
-  '/api/movies/:id',
-  /*checkJwt,*/ async (req, res) => {
-    try {
-      const movie = await Movie.findOneAndUpdate({ _id: req.params.id }, req.body, { new: true })
-      if (!movie) {
-        return res.status(404).send()
-      }
-      res.send(movie)
-    } catch (error) {
-      res.status(400).send({ error: error.message })
+router.patch('/api/movies/:id', checkJwt, async (req, res) => {
+  try {
+    const userId = req.auth.sub
+    const movie = await Movie.findOneAndUpdate({ movieId: req.params.id, userId }, req.body, {
+      new: true,
+    })
+    if (!movie) {
+      return res.status(404).send()
     }
-  },
-)
+    res.send(movie)
+  } catch (error) {
+    res.status(400).send({ error: error.message })
+  }
+})
 
 /**
  * @swagger
@@ -101,7 +121,7 @@ router.patch(
  *       - in: path
  *         name: id
  *         schema:
- *           type: string
+ *           type: number
  *         required: true
  *         description: The movie ID
  *     responses:
@@ -110,22 +130,18 @@ router.patch(
  *       404:
  *         description: Movie not found
  */
-router.delete(
-  '/api/movies/:id',
-  /*checkJwt,*/ async (req, res) => {
-    try {
-      const movie = await Movie.findOneAndDelete({
-        _id: req.params.id,
-      })
-      if (!movie) {
-        return res.status(404).send()
-      }
-      res.send({ message: `Movie ${movie.title} deleted` })
-    } catch (error) {
-      res.status(500).send(error)
+router.delete('/api/movies/:id', checkJwt, async (req, res) => {
+  try {
+    const userId = req.auth.sub
+    const movie = await Movie.findOneAndDelete({ movieId: req.params.id, userId })
+    if (!movie) {
+      return res.status(404).send()
     }
-  },
-)
+    res.send({ message: `Movie ${movie.title} deleted` })
+  } catch (error) {
+    res.status(500).send(error)
+  }
+})
 
 /**
  * @swagger
@@ -138,16 +154,47 @@ router.delete(
  *       500:
  *         description: Internal server error
  */
-router.get(
-  '/api/movies',
-  /*checkJwt,*/ async (req, res) => {
-    try {
-      const movies = await Movie.find({})
-      res.send(JSON.stringify(movies, null, 2))
-    } catch (error) {
-      res.status(500).send(error)
+router.get('/api/movies', checkJwt, async (req, res) => {
+  try {
+    const userId = req.auth.sub
+    const movies = await Movie.find({ userId })
+    res.send(JSON.stringify(movies, null, 2))
+  } catch (error) {
+    res.status(500).send(error)
+  }
+})
+
+/**
+ * @swagger
+ * /api/movies/{movieId}:
+ *   get:
+ *     summary: Retrieves a movie by movieId and userId
+ *     parameters:
+ *       - in: path
+ *         name: movieId
+ *         schema:
+ *           type: number
+ *         required: true
+ *         description: The movie ID
+ *     responses:
+ *       200:
+ *         description: Returns the movie
+ *       404:
+ *         description: Movie not found
+ *       500:
+ *         description: Internal server error
+ */
+router.get('/api/movies/:movieId', checkJwt, async (req, res) => {
+  try {
+    const userId = req.auth.sub
+    const movie = await Movie.findOne({ movieId: req.params.movieId, userId })
+    if (!movie) {
+      return res.status(404).send({ message: 'Movie not found' })
     }
-  },
-)
+    res.send(movie)
+  } catch (error) {
+    res.status(500).send({ error: error.message })
+  }
+})
 
 export default router
